+++ src/code/RepositorySettings.cs
@@ -7,8 +7,11 @@ using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Management.Automation;
+using System.Security.Cryptography;
+using System.Text;
 using System.Xml;
 using System.Xml.Linq;
+using static Microsoft.PowerShell.PSResourceGet.UtilClasses.PSRepositoryInfo;
 
 namespace Microsoft.PowerShell.PSResourceGet.UtilClasses
 {
