+++ src/code/ServerFactory.cs
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Microsoft.PowerShell.PSResourceGet.Cmdlets;
 using Microsoft.PowerShell.PSResourceGet.UtilClasses;
 using System.Management.Automation;
-using System.Management.Automation.Runspaces;
 using System.Net;
 
 namespace Microsoft.PowerShell.PSResourceGet
